Server side clean up, renaming, etc.

+++ api/Metrix.Core/Source/Domain/Measurement.cs
@@ -4,7 +4,7 @@ public class Measurement
 {
   public string MetricKey { get; set; }
 
-  public long Value { get; set; }
+  public double Value { get; set; }
 
   public string Notes { get; set; }
